ajout de la bannier de pub

+++ database/migrations/2021_07_28_191440_create_banniers_table.php
@@ -19,7 +19,7 @@ class CreateBanniersTable extends Migration
             $table->string('image');
             $table->timestamps();
 
-            $table->unsignedBigInteger('id_client');
+            $table->integer('id_client');
             $table->foreign('id_client')->references('id')->on('client_banners');
         });
     }
